collect 2 tests modules into one
